Make filter_vecs a class.
