+++ filter_vecs.cc
@@ -19,31 +19,6 @@
 
  */
 
-#include "defs.h"
-#include "arcdist.h"
-#include "bend.h"
-#include "discard.h"
-#include "duplicate.h"
-#include "filterdefs.h"
-#include "filter.h"
-#include "height.h"
-#include "interpolate.h"
-#include "nukedata.h"
-#include "polygon.h"
-#include "position.h"
-#include "radius.h"
-#include "reverse_route.h"
-#include "smplrout.h"
-#include "sort.h"
-#include "stackfilter.h"
-#include "swapdata.h"
-#include "trackfilter.h"
-#include "transform.h"
-#include "validate.h"
-#include "gbversion.h"
-#include "inifile.h"
-#include "vecs.h"
-
 #include <QtCore/QByteArray>   // for QByteArray
 #include <QtCore/QString>      // for QString
 #include <QtCore/QStringList>  // for QStringList
@@ -55,137 +30,15 @@
 #include <cassert>             // for assert
 #include <cstdio>              // for printf
 
-
-struct fl_vecs_t {
-  Filter* vec;
-  QString name;
-  QString desc;
-};
-
-ArcDistanceFilter arcdist;
-BendFilter bend;
-DiscardFilter discard;
-DuplicateFilter duplicate;
-HeightFilter height;
-InterpolateFilter interpolate;
-NukeDataFilter nukedata;
-PolygonFilter polygon;
-PositionFilter position;
-RadiusFilter radius;
-ReverseRouteFilter reverse_route;
-SimplifyRouteFilter routesimple;
-SortFilter sort;
-StackFilter stackfilt;
-SwapDataFilter swapdata;
-TrackFilter trackfilter;
-TransformFilter transform;
-ValidateFilter validate;
+#include "defs.h"
+#include "filter_vecs.h"
+#include "filter.h"            // for Filter
+#include "gbversion.h"         // for WEB_DOC_DIR
+#include "inifile.h"           // for inifile_readstr
+#include "vecs.h"              // for Vecs
 
 
-static
-const QVector<fl_vecs_t> filter_vec_list = {
-#if FILTERS_ENABLED
-  {
-    &arcdist,
-    "arc",
-    "Include Only Points Within Distance of Arc",
-  },
-  {
-    &bend,
-    "bend",
-    "Add points before and after bends in routes"
-  },
-  {
-    &discard,
-    "discard",
-    "Remove unreliable points with high hdop or vdop"
-  },
-  {
-    &duplicate,
-    "duplicate",
-    "Remove Duplicates",
-  },
-  {
-    &interpolate,
-    "interpolate",
-    "Interpolate between trackpoints"
-  },
-  {
-    &nukedata,
-    "nuketypes",
-    "Remove all waypoints, tracks, or routes"
-  },
-  {
-    &polygon,
-    "polygon",
-    "Include Only Points Inside Polygon",
-  },
-  {
-    &position,
-    "position",
-    "Remove Points Within Distance",
-  },
-  {
-    &radius,
-    "radius",
-    "Include Only Points Within Radius",
-  },
-  {
-    &routesimple,
-    "simplify",
-    "Simplify routes",
-  },
-  {
-    &sort,
-    "sort",
-    "Rearrange waypoints, routes and/or tracks by resorting",
-  },
-  {
-    &stackfilt,
-    "stack",
-    "Save and restore waypoint lists"
-  },
-  {
-    &reverse_route,
-    "reverse",
-    "Reverse stops within routes",
-  },
-  {
-    &trackfilter,
-    "track",
-    "Manipulate track lists"
-  },
-  {
-    &transform,
-    "transform",
-    "Transform waypoints into a route, tracks into routes, ..."
-  },
-  {
-    &height,
-    "height",
-    "Manipulate altitudes"
-  },
-  {
-    &swapdata,
-    "swap",
-    "Swap latitude and longitude of all loaded points"
-  },
-  {
-    &validate,
-    "validate",
-    "Validate internal data structures"
-  }
-#elif defined (MINIMAL_FILTERS)
-  {
-    &trackfilter,
-    "track",
-    "Manipulate track lists"
-  }
-#endif
-};
-
-Filter*
-find_filter_vec(const QString& vecname)
+Filter* FilterVecs::find_filter_vec(const QString& vecname)
 {
   QStringList options = vecname.split(',');
   if (options.isEmpty()) {
@@ -241,8 +94,7 @@ find_filter_vec(const QString& vecname)
   return nullptr;
 }
 
-void
-free_filter_vec(Filter* filter)
+void FilterVecs::free_filter_vec(Filter* filter)
 {
   QVector<arglist_t>* args = filter->get_args();
 
@@ -257,8 +109,7 @@ free_filter_vec(Filter* filter)
   }
 }
 
-void
-init_filter_vecs()
+void FilterVecs::init_filter_vecs()
 {
   for (const auto& vec : filter_vec_list) {
     QVector<arglist_t>* args = vec.vec->get_args();
@@ -271,8 +122,7 @@ init_filter_vecs()
   }
 }
 
-void
-exit_filter_vecs()
+void FilterVecs::exit_filter_vecs()
 {
   for (const auto& vec : filter_vec_list) {
     (vec.vec->exit)();
@@ -283,8 +133,7 @@ exit_filter_vecs()
  *  Display the available formats in a format that's easy for humans to
  *  parse for help on available command line options.
  */
-void
-disp_filter_vecs()
+void FilterVecs::disp_filter_vecs() const
 {
   for (const auto& vec : filter_vec_list) {
     printf("	%-20.20s  %-50.50s\n",
@@ -292,17 +141,17 @@ disp_filter_vecs()
     const QVector<arglist_t>* args = vec.vec->get_args();
     if (args) {
       for (const auto& arg : *args) {
-        if (!(arg.argtype & ARGTYPE_HIDDEN))
+        if (!(arg.argtype & ARGTYPE_HIDDEN)) {
           printf("	  %-18.18s    %-.50s %s\n",
                  arg.argstring, arg.helpstring,
                  (arg.argtype & ARGTYPE_REQUIRED) ? "(required)" : "");
+        }
       }
     }
   }
 }
 
-void
-disp_filter_vec(const QString& vecname)
+void FilterVecs::disp_filter_vec(const QString& vecname) const
 {
   for (const auto& vec : filter_vec_list) {
     if (vecname.compare(vec.name, Qt::CaseInsensitive) != 0) {
@@ -313,17 +162,17 @@ disp_filter_vec(const QString& vecname)
     const QVector<arglist_t>* args = vec.vec->get_args();
     if (args) {
       for (const auto& arg : *args) {
-        if (!(arg.argtype & ARGTYPE_HIDDEN))
+        if (!(arg.argtype & ARGTYPE_HIDDEN)) {
           printf("	  %-18.18s    %-.50s %s\n",
                  arg.argstring, arg.helpstring,
                  (arg.argtype & ARGTYPE_REQUIRED) ? "(required)" : "");
+        }
       }
     }
   }
 }
 
-static
-void disp_help_url(const fl_vecs_t& vec, const arglist_t* arg)
+void FilterVecs::disp_help_url(const fl_vecs_t& vec, const arglist_t* arg)
 {
   printf("\t" WEB_DOC_DIR "/fmt_%s.html", CSTR(vec.name));
   if (arg) {
@@ -331,8 +180,7 @@ void disp_help_url(const fl_vecs_t& vec, const arglist_t* arg)
   }
 }
 
-static void
-disp_v1(const fl_vecs_t& vec)
+void FilterVecs::disp_v1(const fl_vecs_t& vec)
 {
   disp_help_url(vec, nullptr);
   printf("\n");
@@ -360,8 +208,7 @@ disp_v1(const fl_vecs_t& vec)
  *  parse.   Typically invoked by programs like graphical wrappers to
  *  determine what formats are supported.
  */
-void
-disp_filters(int version)
+void FilterVecs::disp_filters(int version) const
 {
   auto sorted_filter_vec_list = filter_vec_list;
 
@@ -388,15 +235,14 @@ disp_filters(int version)
   }
 }
 
-static bool
-validate_filter_vec(const fl_vecs_t& vec)
+bool FilterVecs::validate_filter_vec(const fl_vecs_t& vec)
 {
   bool ok = Vecs::validate_args(vec.name, vec.vec->get_args());
 
   return ok;
 }
 
-bool validate_filters()
+bool FilterVecs::validate_filters() const
 {
   bool ok = true;
